Make ML libraries optional, fix type issues
